+++ static/app/views/preprod/types/installDetailsTypes.ts
@@ -1,0 +1,9 @@
+import type {Platform} from './sharedTypes';
+
+export interface InstallDetailsApiResponse {
+  platform: Platform;
+  codesigning_type?: string;
+  install_url?: string;
+  is_code_signature_valid?: boolean;
+  profile_name?: string;
+}
