+++ static/app/views/preprod/types/sharedTypes.ts
@@ -1,0 +1,1 @@
+export type Platform = 'ios' | 'android' | 'macos';
